Restructured the overlay config
