+++ src/main/java/io/github/chindeaone/collectiontracker/config/categories/Overlay.java
@@ -1,0 +1,29 @@
+package io.github.chindeaone.collectiontracker.config.categories;
+
+import com.google.gson.annotations.Expose;
+import io.github.chindeaone.collectiontracker.config.categories.overlay.OverlayList;
+import io.github.chindeaone.collectiontracker.config.categories.overlay.OverlaySingle;
+import io.github.moulberry.moulconfig.annotations.Accordion;
+import io.github.moulberry.moulconfig.annotations.ConfigEditorBoolean;
+import io.github.moulberry.moulconfig.annotations.ConfigOption;
+
+public class Overlay {
+
+    @Expose
+    @ConfigOption(name = "Simple Overlay", desc = "")
+    @Accordion
+    public OverlaySingle overlaySingle = new OverlaySingle();
+
+    @Expose
+    @ConfigOption(name = "Overlay List", desc = "Choose between multiple collections to track.")
+    @Accordion
+    public OverlayList overlayList = new OverlayList();
+
+    @Expose
+    @ConfigOption(
+            name = "Overlay Text Color",
+            desc = "Toggle this to enable color-coded text."
+    )
+    @ConfigEditorBoolean
+    public boolean overlayTextColor = false;
+}
